+++ Q3E/src/main/jni/doom3/neo/externlibs/curl/lib/mprintf.c
@@ -864,7 +864,7 @@ static int dprintf_formatf(
 	  base = 16;
 	  digits = (p->flags & FLAGS_UPPER)? upper_digits : lower_digits;
 	  alt = 1;
-	  num = (unsigned long) ptr;
+	  num = (/*unsigned long //k*/uintptr_t) ptr;
 	  is_neg = 0;
 	  goto number;
 	}
